requirements.txt
affine==2.3.0
altair==4.1.0
anyio==3.0.1
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bleach==3.3.0
bokeh==2.3.1
Bottleneck==1.3.2
branca==0.4.2
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cloudpickle==1.6.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
cytoolz==0.11.0
dask==2021.4.1
decorator==5.0.7
defusedxml==0.7.1
deprecation==2.1.0
Django==3.2.1
django-bootstrap3==15.0.0
django-extensions==3.1.3
django-smart-selects==1.5.9
django-storages==1.8
django-tinymce==3.3.0
dj-database-url==0.5.0
dj-static==0.0.6
earthpy==0.9.2
entrypoints==0.3
Fiona==1.8.18
folium==0.0.0
fsspec==2021.4.0
future==0.18.2
GDAL==3.2.1
GeoAlchemy2==0.8.5
geocoder==1.38.1
geopandas==0.9.0
greenlet==1.0.0
gunicorn==20.0.4
htmlmin==0.1.12
hydrofunctions==0.1.7
idna==2.10
imagecodecs==2021.3.31
ImageHash==4.2.0
imageio==2.9.0
importlib-metadata==4.0.1
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.0
jupyter-packaging==0.9.2
jupyter-server==1.6.4
kiwisolver==1.3.1
locket==0.2.0
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
missingno==0.4.2
mistune==0.8.4
multimethod==1.4
multitasking==0.0.9
munch==2.5.0
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.3.0
numpy==1.20.2
odfpy==1.4.1
olefile==0.46
packaging==20.9
pandas==1.2.4
pandas-profiling==3.0.0
pandocfilters==1.4.2
parso==0.8.2
partd==1.2.0
patsy==0.5.1
phik==0.11.2
pickleshare==0.7.5
Pillow==8.1.2
pip==21.1.1
pooch==1.3.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
pydantic==1.8.2
pygeos==0.9
Pygments==2.8.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyproj==3.0.1
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
PySocks==1.7.1
pystache==0.5.4
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==1.0.1
PyYAML==5.4.1
pyzmq==22.0.3
rasterio==1.2.2
ratelim==0.1.6
requests==2.25.1
Rtree==0.9.7
scikit-image==0.18.1
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
setuptools==49.6.0.post20210108
Shapely==1.7.1
six==1.15.0
sniffio==1.2.0
snuggs==1.4.7
sorl-thumbnail==12.7.0
SQLAlchemy==1.4.13
sqlparse==0.4.1
static==1.1.1
static3==0.7.0
statsmodels==0.12.2
tangled-up-in-unicode==0.1.0
terminado==0.9.4
testpath==0.4.4
tifffile==2021.4.8
tomlkit==0.7.0
toolz==0.11.1
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
vega-datasets==0.9.0
visions==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
wincertstore==0.2
win-inet-pton==1.1.0
yfinance==0.1.59
zipp==3.4.1
